ref(overlay): Rename `fullscreen` option to `openOnInit`
Previously, the name was missleading, suggesting it'd make open the overlay
in fullscreen rather than windowed mode, which it never did.
The only thing this option does is controlling if the overlay window
should be opened immediately when the UI is rendered or if just the
button should be displayed.