fix: Fix ups for the vite-plugin PR
Fixes up #434.

We sometimes load Spotlight befor a Sentry SDK is loaded which
then bails on running the integration with it, making it lose
most of its appeal. To fix this, this PR adds a retry mechanism
which tries detecting Sentry SDK up to 3 times.

This works on the first try if it's just an ordering problem. It
however would not work where a Sentry SDK is loaded on-demand as
we don't have an "on load event" for Sentry SDK.