feat: Move contextlines code to sidecar (#454)
This change moves the contextlines provider to the sidecar from the Vite
plugin opening it up to all SDKs and integrations, not just Vite-based
ones. This also means we do not clobber Vite dev server's URL-space (not
that it was a problem, but alas).

This required adding a setter to `SentryDataCache` class as it was used
as a singleton from its module and handles the additional context line
fetching per earlier PRs. It would be good to refactor this usage to
make it more contained/encapsulated.

Closes #449.