fix: Fix npx @spotlightjs/spotlight
Due to a binary name clash between @spotlightjs/spotlight and @spotlightjs/sidecar (which is a dependency of the former), `npx @spotlightjs/spotlight` was running the binary from the sidecar, which does not have the standalone overlay. This PR fixes the issue by renaming the `@spotlightjs/spotlight` binary name to `spotlight` (from `spotlight-sidecar`).

Fixes #385.