feat: Add 'open in editor' icon (#462)
Closes #459.

Adds a little pencil icon that opens the file in the editor when
clicked. This is done on the sidecar side.

PS: Need to document `LAUNCH_EDITOR` env variable. We use `EDITOR` by
default and if it is `vim` or something terminal based, it is opened on
the terminal that runs the sidecar, which becomes confusing at times.

![image](https://github.com/user-attachments/assets/af701484-6adf-4e93-ab34-10502d85728d)