ci: Fix auto publishing to Docker & Electron
As seen from #484, we don't automatically publish to Docker and Electron targets. The reason for this is we have an strict `inputs.docker` and `inputs.electron` check, assuming they will have a default value of `true`. That said the `input` context only gets populated on a manually triggered workflow, hence these jobs get skipped on auto publish actions. This PR should fix the issue by replacing it with an explicit `!= false` check, so it becomes `true` if the value is `null` or `undefined`.