ref: Overhaul console integration (#525)
Our console integration was designed to work on the same page with the
Spotlight UI. It was using a naive filtering method to avoid infinite
loops from Spotlight debug mode logger itself. Moreover, it was not
taking the `sidecarURL` option, was defaulting to `localhost:8969` as
the sidecar URL. This patch overhauls the two, leveraging the fact that
they are on the same page. This means switching to the local event-based
message passing between the UI and the console integration, removing the
need for `fetch` calls and passing around the sidecarURL option. It also
overhauls the Spotlight logger to always try to use an unmodified
console implementation, removing the need for hacky filtering on the
console integration side.

ref #521.