feat: Add startFrom init option
Adds a `startFrom` init option to control the starting path of Spotlight app. This is useful to immediately display an error when we know will be there. The patch also fixes a bunch of init logic where `__spotlight.initOptions` were not being respected when custom init options are used. It also fixes `sentry()` integration getting initialized even if `integrations` did not include it.