feat(spotlight): Add self-contained spotlight binary creation (#559)
This PR adds a new release type which is a self-contained CLI akin to
running `npx @spotlightjs/spotlight` but without any `node` or `npx`
or any other requirements on the system.

It also changes how we serve the static assets: instead of serving the
entire folder, which can be dangerous as it allows access to arbitrary
files, we now use the `manifest.json` generated when compiling it and
only serve files listed there, directly from memory. This should also
increase the performance.