fix(spotlight): Fix UI not working with different port take 2 (#623)
With the previous patch, we assumed the page loads `assets/main.js` is
_the_ Spotlight main page which is not the case for auto injection by
SDKs (only Django for now but this should expand). This patch fixes the
issue by injecting `sentrySpotlightPort` in server-timing headers (yes,
hack!)to avoid dynamic HTML generation from sidecar side. I tested that
it works and now that it works, we may use this from Django too!