fix(overlay): Make envelope parsing resillient against bad data
Fixes #628. Looks like we send incorrect payload length for certain replay envelopes, causing Spotlight to completely crash when trying to parse them as it tries to read the remaining binary data (after the incorrectly short payload length) as JSON header for the next payload and fail. This patch simply skips the rest of the envelope once a header reading error happens.