fix: Binary to UTF-8 encoding shenanigans
We've been corrupting binary envelopes both on Sidecar and Overlay due to implicit utf-8 encodings all around. This patch fixes #628 for reals by keeping envelope data as binary until it reaches the SSE transport, and then uses a weird byte-by-byte conversion to Uint8Array on client side from the UTF-8 encoded version instead of using the `TextEncoder()` (as it also uses utf-8 instead of binary. The ideal solution to this would be to simply encode envelopes with `base64` but that seems like a breaking change, so postponing for now with proper comments in the code.