feat(overlay): Trace details layout change (#831)
## Layout and minor changes:
- Now trace title shows up at the top when you click on a trace. Because
of this, now there's not Trace header on the trace details view as the
info is already at the top.
- Traces that contain AI now  have an AI tag for better discoverability.


https://github.com/user-attachments/assets/9458148f-1626-4eb2-85f1-a56632620f47

### Notes
- Now there's no X button on trace details. Not sure if that's a good
idea. Now I'm used to closing the trace details by clicking on the trace
header but may no be intuitive for newbies.


Before opening this PR:

- [x] I added a [Changeset
Entry](https://spotlightjs.com/contribute/changesets/) with `pnpm
changeset:add`
- [ ] I referenced issues that this PR addresses