fix: electron app startup issue (#923)
The issue was occurring because of 2 bugs -

1. The main issue, `afterAllSetup` handler function, this sends a
serialized envelope to the Spotlight frontend. Even tho we are not using
this envelope it is getting passed through the `processEnvelope`
function. Now because we are handling the parsing/processing of
envelopes on the backend, this was failing.
2. Because of the above issue, the electron app was trying to run
`showErrorMessage` before showing the "askForPermission" dialog which
was also failing.

Due to these reason the app was unable to startup.

Also some minor improvements I did while figuring out the issues
- [x] Added types for electron's contextBridge
- [x] Deleted props for the old Spotlight Integration
- [x] Changed Event to Envelope (Performance improvement)