feat(overlay): Updated Routing in overlay and removed sentry-integration (#921)
- Simplified the App component by utilizing React Router for navigation.
- Adjusted links and navigation to reflect the new routing structure.

This refactor aims to improve code maintainability and clarity in the
telemetry section of the overlay.

<!--
Tick these boxes if they're applicable to your PR.
- Changesets are only required for PRs to Spotlight library packages
(e.g. @spotlightjs/overlay). Not for the website/docs or demo app
contributions.
- Typo correction or small bugfix PRs don't require an issue. If you're
making a bigger change, please open an issue first.
-->

Before opening this PR:

- [x] I added a [Changeset
Entry](https://spotlightjs.com/contribute/changesets/) with `pnpm
changeset:add`
- [x] I referenced issues that this PR addresses
#891 

- Updated Router in Overlay - BrowserRouter for url-based routing
- prefixed telemetry routes with /telemetry (default)
- removed sentry-integration -> this was useful in case of embedding.
But not useful in the case of electron and standalone.
- In case of Electron - this was tracking the telemetry data from Sentry
Electron SDK used in the Spotlight Electron App.

---------

Co-authored-by: Burak Yigit Kaya <byk@sentry.io>