Add a CLI that agents can also use instead of the MCP (#997)
This CLI introduces commands like `logs` or `errors` which can also be
combined with a `,` or `+` to listen to multiple streams at once.

The first iteration will be a streaming log like `tail -f` but in the
next
iteration we will add a `no-follow` mode that accepts some filters like
a
`timeWindow` or `limit` as the MCP does.

Closes #958.

---------

Co-authored-by: mathuraditya724 <mathuraditya724@gmail.com>
Co-authored-by: betegon <miguelbetegongarcia@gmail.com>