feat(spotlight): Use dynamic port for run cmd (#1036)
Closes #1025.

Had to simplify and modularize the functions in `main.ts` -- especially
splitting out the `stdioMCP` logic into the `cli/mcp.ts` where it should
have always been. This made things a lot more simple for the regular
`server` command and all MCP-related complex logic is now contained in
`cli/mcp.ts` and in a more understandable way.

Rest is some more reorganization, and essentially making sure we return
the created server instance from `server` and `tail` commands so we can
get the automatically assigned port from there when a dynamic port is
used.