feat(sidecar): Auto capture stdio logs as Sentry log events
When using `spotlight run` if the process is not using Sentry SDK
with logs integration, we lose all its logging output that comes
to stdout/stderr. This change automatically captures that output
checks against any log items we receive from the process on the
sidecar and sends them as Sentry log events only if we cannot
get any logs from our integration.