fix(sidecar): Fix disconnect every 1s
After #1059 which set a 1s timeout on the server, our SSE connection kept getting dropped after every second. This PR fixes the issue by removing that and re-introducing `server.closeAllConnections()` call in the shutdown routine. The reason we could not use this was because `ServerType` from Hono covered both HTTP/1 and HTTP/2 servers where the HTTP/2 version lacked this function. Spotlight will be HTTP/1 for the foreseable future so this patch casts the server to Node's native `Server` type and uses the `closeAllConnections()` method which allows graceful shutdown.