Fix(Electron): Electron window dragging and traffic light positioning on macOS (#1173)
We had two issues  on macOS Electron app:

- Window not draggable
- The close/minimize/maximize buttons were positioned on top of the
Spotlight logo in the sidebar

We were checking `IN_DESKTOP_ENV` which was never set anywhere, so the
conditional `mt-8` margin were we set the draggable area and the
semaphore was place wasn't applied.

I believe this was caused when we merged all packages into the spotlight
one.



https://github.com/user-attachments/assets/ceedd707-7c14-471d-a824-1fe31024a528

---------

Co-authored-by: Burak Yigit Kaya <byk@sentry.io>